style : Update UI components
