+++ components/header.tsx
@@ -5,9 +5,14 @@ import React from "react";
 import logo from "@/assets/logo.png";
 
 function Header() {
+  const navItems = [
+    { href: "/mint", label: "Mint" },
+    { href: "/airdrop", label: "Airdrop" },
+  ];
+
   return (
     <>
-      <section className="absolute top-0 z-[100] flex items-center justify-between w-full px-10 ">
+      <section className=" absolute top-0 z-[100] flex items-center justify-between w-full px-10 py-1 ">
         <div className="flex">
           <Image
             className="cursor-pointer"
@@ -17,27 +22,24 @@ function Header() {
             alt="logo"
           />
           <Link href="/">
-            <div className="mt-[3vh] text-pink-500 mb-2.5  -top-9 font-bold text-xl  md:text-xl lg:text-xl lg:font-semibold  dark:text-zinc-400 max-w-4xl leading-relaxed lg:leading-snug text-left mx-auto ">
+            <div className="mt-[3vh] text-primary -top-9 font-bold text-xl md:text-xl lg:text-xl lg:font-semibold  max-w-4xl leading-relaxed lg:leading-snug text-left mx-auto">
               DragonSoulBoundNFT
             </div>
           </Link>
         </div>
 
-        <div className="flex">
-          <ul className="mt-[4vh] flex space-x-4 mb-2.5 -top-9 text-xl px-4 md:text-xl lg:text-xl font-semibold text-neutral-700 dark:text-zinc-400 max-w-4xl leading-relaxed lg:leading-snug text-left mx-auto">
-            <li>
-              <Link href="/mint">
-                <div className="text-red-500">Mint</div>
-              </Link>
-            </li>
-            <li>
-              <Link href="/airdrop">
-                <div className="text-green-500">Airdrop</div>
-              </Link>
-            </li>
+        <div className="flex gap-10">
+          <ul className="flex gap-10  text-xl px-4 md:text-xl lg:text-xl font-semibold text-neutral-700 dark:text-zinc-400 max-w-4xl leading-relaxed lg:leading-snug text-left mx-auto">
+            {navItems.map((item, index) => (
+              <li key={index} className="py-1">
+                <Link href={item.href}>
+                  <div className="text-secondary">{item.label}</div>
+                </Link>
+              </li>
+            ))}
           </ul>
-          <div className="mt-[3vh] connect-btn">
-            <ConnectButton />
+          <div className=" connect-btn">
+            <ConnectButton accountStatus="avatar" />
           </div>
         </div>
       </section>
