+++ components/homepage.tsx
@@ -4,7 +4,6 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 import { motion } from "framer-motion";
-
 import { useToast } from "@/components/ui/use-toast";
 import {
   type BaseError,
@@ -18,6 +17,8 @@ import { useChainId } from "wagmi";
 import { CONTRACT_ADDRESS_BAOBAB, CONTRACT_ADDRESS_CYPRESS } from "./contract";
 import { Button } from "./ui/button";
 import Link from "next/link";
+import HeroImage from "./svgcomponents/HeroImage";
+// import HeroImage from "@/assets/HeroImage.svg";
 
 const formSchema = z.object({
   to: z.coerce.string({
@@ -77,45 +78,51 @@ function HomePage() {
   return (
     <div className="w-full">
       <div>
-        <Hero className="px-10">
-          <motion.h1
-            initial={{
-              opacity: 0,
-              y: 20,
-            }}
-            animate={{
-              opacity: 1,
-              y: [20, -5, 0],
-            }}
-            transition={{
-              duration: 0.5,
-              ease: [0.4, 0.0, 0.2, 1],
-            }}
-            className="text-xl md:text-xl lg:text-xl font-semibold text-neutral-700 dark:text-zinc-400 max-w-4xl leading-relaxed lg:leading-snug text-left"
-          >
-            <Highlight className="mb-2.5 text-7xl -top-9 font-bold">
-              Mint SoulBound NFT
-            </Highlight>
-            {/* break line */} <br />
-            The{" "}
-            <span className="bg-gradient-to-r from-sky-400 to-blue-600 bg-clip-text text-transparent">
-              SoulBoundNFT
-            </span>{" "}
-            minter dapp will leverage the robust infrastructure of the Klaytn
-            blockchain, renowned for its scalability, security, and
-            developer-friendly environment. Through this dapp, users will have
-            the power to immortalize their digital creations, whether it be
-            artwork, music, or any other form of digital content, as
-            SoulBoundNFTs, imbued with a sense of authenticity and exclusivity.
-          </motion.h1>
-          <Button
-            variant="outline"
-            size="default"
-            className="bg-gradient-to-r from-sky-400 to-blue-600 flex mt-6   text-white">
-            <Link href="/mint">Mint SoulBound NFT Now</Link>
-          </Button>
+        <Hero className="flex items-center px-10">
+          <div>
+            <motion.h1
+              initial={{
+                opacity: 0,
+                y: 20,
+              }}
+              animate={{
+                opacity: 1,
+                y: [20, -5, 0],
+              }}
+              transition={{
+                duration: 0.5,
+                ease: [0.4, 0.0, 0.2, 1],
+              }}
+              className="text-xl md:text-xl lg:text-xl font-semibold text-white  dark:text-zinc-400 max-w-4xl leading-relaxed lg:leading-snug text-left"
+            >
+              <Highlight className="mb-2.5 text-7xl -top-9 font-bold">
+                Mint SoulBound NFT
+              </Highlight>
+              {/* break line */} <br />
+              The{" "}
+              <span className="bg-primary bg-clip-text text-transparent">
+                SoulBoundNFT
+              </span>{" "}
+              minter dapp will leverage the robust infrastructure of the Klaytn
+              blockchain, renowned for its scalability, security, and
+              developer-friendly environment. Through this dapp, users will have
+              the power to immortalize their digital creations, whether it be
+              artwork, music, or any other form of digital content, as
+              SoulBoundNFTs, imbued with a sense of authenticity and exclusivity.
+            </motion.h1>
+            <Link href="/mint">
+              <Button
+                variant="default"
+                size="default"
+                className="bg-primary flex mt-6   text-white">
+                Mint SoulBound NFT Now
+              </Button>
+            </Link>
+          </div>
+          <div className="w-[40%]">
+            <HeroImage />
+          </div>
         </Hero>
-        
       </div>
     </div>
   );
